util: move server.TokenLimiter into util
