+++ pkg/util/tokenlimiter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package util
 
 // Token is used as a permission to keep on running.
 type Token struct {
